CompoundV3ERC4626 vault
Permissioned ERC4626 that invests into Compound. Also a user with
HARVEST_ROLE can claim the rewards that would be swapped for the asset
and supplied again to Compound (accounting then as a profit for all the
LPs).

Also did a refactoring in SharedSmartVault, moving the permissions to
PermissionedERC4626.

Code should be complete, but no real tests have been made.