CompoundV3ERC4626 tests and change in PermissionedERC4626
Previously PermissionedERC4626 was checking by the role of the calling
user and reimplementing the deposit/withdraw/mint/redeem methods. Now
leaves those methods unchanged and if the user doesn't have LP_ROLE
returns 0 on maxDeposit and maxMint.

This means that now anyone can enter or exit, but the owner of the
shares needs to be whitelisted (LP_ROLE).

Also, anyone with shares can exit.