Pluggable strategies with Delegatecall
New approach to vaults. The idea is to have the vault specific code into
a single contract (permissions and other stuff) and implement the
specific investments as "strategy" contracts that only have the logic to
deposit or withdraw from a specific investment.

So far only implemented CompoundV3InvestStrategy (not tested yet).

This will make sense when in future PRs we implement a multi-strategy
vault, that deploys funds to several underlying protocols (AAVE,
Compound, etc.)