Big refactoring: removed storageSlot as parameter
The storageSlot is calculated on each strategy on deployment and then is
fixed and doesn't need to be passed from the vault to the strategy.

Also, I added view methods in the library to automate the passing of
`address(this)` as a first parameter for the views.

After all this, the usage of the strategies feels very natural as if
they were actual contracts.